refactor: rename AddSurveyModel to Survey and move it to domain models
