+++ src/data/usecases/add-survey/db-add-survey-usecase-protocols.ts
@@ -1,2 +1,3 @@
 export * from '@src/data/protocols/db'
 export * from '@src/domain/usecases/add-survey'
+export * from '@src/domain/models/survey'
