+++ src/data/usecases/add-survey/db-add-survey-usecase.test.ts
@@ -1,10 +1,12 @@
-import { AddSurveyModel } from '@src/domain/usecases/add-survey'
-import { AddSurveyRepository } from './db-add-survey-usecase-protocols'
+import {
+  AddSurveyRepository,
+  SurveyModel
+} from './db-add-survey-usecase-protocols'
 import { AddSurveyRepositoryError } from './db-add-survey-usecase-result'
 import { DbAddSurveyUseCase } from './db-add-survey-usecase'
 
 class AddSurveyRepositoryStub implements AddSurveyRepository {
-  async add(_survey: AddSurveyModel): Promise<void> {
+  async add(_survey: SurveyModel): Promise<void> {
     return undefined
   }
 }
