+++ src/data/usecases/add-survey/db-add-survey-usecase.ts
@@ -1,6 +1,6 @@
 import { failure, success } from '@src/shared'
 import {
-  AddSurveyModel,
+  SurveyModel,
   AddSurveyRepository,
   AddSurveyUseCase
 } from './db-add-survey-usecase-protocols'
@@ -12,7 +12,7 @@ import {
 export class DbAddSurveyUseCase implements AddSurveyUseCase<DbAddSurveyResult> {
   constructor(private readonly addSurveyRepository: AddSurveyRepository) {}
 
-  async add(survey: AddSurveyModel): Promise<DbAddSurveyResult> {
+  async add(survey: SurveyModel): Promise<DbAddSurveyResult> {
     try {
       await this.addSurveyRepository.add(survey)
       return success(undefined)
