+++ src/domain/models/survey.ts
@@ -1,0 +1,10 @@
+export interface SurveyAnswer {
+  image: string
+  answer: string
+}
+
+export interface SurveyModel {
+  question: string
+  answers: SurveyAnswer[]
+  date: Date
+}
