+++ src/domain/usecases/survey/index.ts
@@ -1,0 +1,1 @@
+export * from './load'
