+++ src/infra/db/mongodb/survey/survey-mongo-repository.ts
@@ -1,9 +1,9 @@
 import { AddSurveyRepository } from '@src/data/protocols/db'
-import { AddSurveyModel } from '@src/domain/usecases/add-survey'
+import { SurveyModel } from '@src/domain/models/survey'
 import { MongoHelper } from '@src/infra/db/mongodb/helpers/mongo-helper'
 
 export class SurveyMongoRepository implements AddSurveyRepository {
-  async add(survey: AddSurveyModel): Promise<void> {
+  async add(survey: SurveyModel): Promise<void> {
     const surveyCollection = await MongoHelper.getCollection('surveys')
     await surveyCollection.insertOne(survey)
   }
